requirements-automation.txt
# Requirements for Documentation Pipeline Automation
# Install with: pip install -r requirements-automation.txt

# Core dependencies
pyyaml>=6.0
requests>=2.28.0

# AI integration
anthropic>=0.8.0

# Optional: For semantic analysis
# scikit-learn>=1.3.0
# numpy>=1.24.0

# Optional: For search indexing
# beautifulsoup4>=4.12.0
# pinecone-client>=2.2.0

# Development
pytest>=7.0.0
black>=23.0.0
